Bug fixing, more stability
Voltage logging is now stable

Altitude errors were fixed; it was interpreting altitude as a float
instead of as a double because in CurrentState.h it was marked as a
float

+++ src/data/CurrentState.h
@@ -65,7 +65,7 @@ class CurrentState {
 		double longitude;
 		float battery_voltage;
 		float battery_remaining;
-		float altitude;
+		double altitude;
 		float ground_speed;
 		float throttle;
 		float dist_to_home;
